Add contributor guidelines and unit tests
