+++ package.json
@@ -5,12 +5,15 @@
   "main": "main.js",
   "scripts": {
     "start": "electron .",
-    "test": "echo \"No tests\""
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "electron": "^37.3.0"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0"
   }
 }
